Cleaned file structure; added a feature to the table
