+++ frontend/src/components/TripComponents/TripContainer.js
@@ -1,15 +1,14 @@
 import React from 'react';
 import HeaderContainer from '../HeaderComponents/HeaderContainer';
 import TripHeader from './TripHeader';
-//import TripTable from './TripTable';
-import TableTest from './TableTest'
+import TripTable from './TripTable'
 
 const TripContainer = () => {
     return (
         <div>
             <HeaderContainer />
             <TripHeader />
-            <TableTest />
+            <TripTable />
         </div>
     )
 };
